backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
cerebras-cloud-sdk